backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0  # For LLM-based content improvements
python-dotenv>=1.0.0  # For environment variable management 